moved the screen size change function into Context, added MobileDateTimePicker
