+++ src/contexts/UIContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const UIContext = createContext(null);
